backend/code/008/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mve-008-cos-log-linear"
version = "0.1.0"
description = "MVE 008: Cosine-Reweighted Log-Linear Attention on MQAR"
requires-python = ">=3.11"

dependencies = [
    "setuptools>=61.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "modal>=1.3.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "data*", "train*"]
